move rollup config into root and move plugins around
